Add class names to li
